% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jatos.R
\name{read_jatos}
\alias{read_jatos}
\title{Read result file from Jatos}
\usage{
read_jatos(result_file, add_unique_ids = TRUE)
}
\arguments{
\item{result_file}{The path to the result text file exported from Jatos.}

\item{add_unique_ids}{Adds a unique ID to each participant. Default TRUE. If false, the `participant` value from Jatos is taken.}
}
\value{
A data frame
}
\description{
This function reads the result text file exported from Jatos and creates a data frame.
By default, it assigns each participant a unique ID.
}
\examples{
data <- read_jatos('myresults.txt')


Possible issue: if the result file contains nested lists, it is likely that they will end up in your data frame too.
Use the tidyverse `unnest_wider` function to expand them as new columns.
Currently, only the nested lists from `survey` items are handled.
}
