% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_counter.R
\name{update_counter}
\alias{update_counter}
\title{Update the Jatos counter for your experiments}
\usage{
update_counter(jatos_path)
}
\arguments{
\item{jatos_path}{The path to your Jatos folder (often called `jatos_XXX_java`).}
}
\description{
This function updates the counter.js file of all your Jatos experiments.
All you have to do is to give the path to your Jatos folder.
}
\examples{
update_counter('/home/user/jatos_mac_java')
update_counter('/home/user/jatos_linux_java')

}
