% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jatos.R
\name{read_jatos}
\alias{read_jatos}
\title{Read result file from Jatos}
\usage{
read_jatos(
  result_file,
  add_unique_ids = FALSE,
  flatten = TRUE,
  remove = "response."
)
}
\arguments{
\item{result_file}{The path to the result text file exported from Jatos.}

\item{add_unique_ids}{Default FALSE. Adds a sequential ID to participants. If true, the `participant` value from Jatos is erased.}

\item{flatten}{Default `TRUE`. automatically `flatten()` nested data frames into a single non-nested data frame. From `jsonlite` package.}
}
\value{
A data frame
}
\description{
This function reads the result text file exported from Jatos and creates a data frame.
It can also reassign each participant a sequential ID.
}
\examples{
data <- read_jatos('myresults.txt')


}
