% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stimuli.R
\name{write_stimuli}
\alias{write_stimuli}
\title{Write stimuli for Jatos}
\usage{
write_stimuli(csv_file, experiment_folder, delim = ",", null_string = TRUE)
}
\arguments{
\item{experiment_folder}{The path to the folder of your experiment.}

\item{delim}{Delimitation symbol used in the .csv file. Default is ",".}

\item{null_string}{Default `TRUE`. `TRUE` includes empty strings for unfilled columns of the stimuli csv file. `FALSE` removes unfilled columns.}

\item{csvfile}{The path to the csv file containing your items.}
}
\description{
This function converts a csv file containing experimental items into a stimuli.js file with JSON format.
The stimuli.js file has to added in the Jatos folder of your experiment (in jatos/study_assets_root).
}
\examples{
write_stimuli('myexperiment/items.csv','/jatosfolder')

}
